requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.0
pydantic==2.5.3
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
python-multipart==0.0.6
jinja2==3.1.3
python-dotenv==1.0.0
